Authentication and Authorization with default admin is almost done

+++ src/main/java/com/example/dbnode/api/client/controller/database/DatabaseController.java
@@ -13,20 +13,24 @@ public class DatabaseController {
     private final DatabaseService databaseService;
     @PostMapping("node/client/create/database/{databaseName}")
     public ResponseEntity<String> createDatabase(@PathVariable("databaseName") String databaseName){
+        if(databaseName.equals("db_system"))
+            return ResponseEntityCreator.getResponse(HttpStatus.FORBIDDEN,"This database name reserved for system");
         boolean isDatabaseCreated = databaseService.createDatabase(databaseName);
         if(isDatabaseCreated)
-            return getResponse(HttpStatus.CREATED,"Database was created successfully!");
+            return ResponseEntityCreator.getResponse(HttpStatus.CREATED,"Database was created successfully!");
         else
-            return getResponse(HttpStatus.CONFLICT,"Database you tried to create is already exist");
+            return ResponseEntityCreator.getResponse(HttpStatus.CONFLICT,"Database you tried to create is already exist");
     }
 
     @DeleteMapping("node/client/delete/database/{databaseName}")
     public ResponseEntity<String> deleteDatabase(@PathVariable("databaseName") String databaseName){
+        if(databaseName.equals("db_system"))
+            return ResponseEntityCreator.getResponse(HttpStatus.FORBIDDEN,"db_system can not be deleted!");
         boolean isDatabaseDeleted = databaseService.deleteDatabase(databaseName);
         if(isDatabaseDeleted)
-            return getResponse(HttpStatus.OK,"Database was deleted successfully!");
+            return ResponseEntityCreator.getResponse(HttpStatus.OK,"Database was deleted successfully!");
         else
-            return getResponse(HttpStatus.NOT_FOUND,"Database you tried to delete does not exist");
+            return ResponseEntityCreator.getResponse(HttpStatus.NOT_FOUND,"Database you tried to delete does not exist");
     }
 
     @GetMapping("node/client/list/databases")
@@ -34,9 +38,4 @@ public class DatabaseController {
         return ResponseEntityCreator.getResponse(HttpStatus.OK,databaseService.listDatabases());
     }
 
-    private ResponseEntity<String> getResponse(HttpStatus status, String body){
-        if(status == null || body == null)
-            throw new IllegalArgumentException();
-        return ResponseEntity.status(status).body(body);
-    }
 }
